moved fragments by packages, add hilt annotation into fragments, add parent class for fragments
