--- app/src/main/java/com/example/fitnessconstructor/AppSettingsFragment.kt
@@ -1,32 +1,0 @@
-package com.example.fitnessconstructor
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-
-
-class AppSettingsFragment : Fragment() {
-
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-    }
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_app_settings, container, false)
-    }
-
-    companion object {
-
-        @JvmStatic
-        fun newInstance() = AppSettingsFragment()
-
-    }
-}
