--- app/src/main/java/com/example/fitnessconstructor/MainActivity.kt
@@ -1,17 +1,0 @@
-package com.example.fitnessconstructor
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import dagger.hilt.android.AndroidEntryPoint
-
-
-@AndroidEntryPoint
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-
-        //remove action bar while launch app
-        supportActionBar?.hide()
-    }
-}
