--- app/src/main/java/com/example/fitnessconstructor/WorkoutListFragment.kt
@@ -1,25 +1,0 @@
-package com.example.fitnessconstructor
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-
-
-class WorkoutListFragment : Fragment() {
-
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_workout_list, container, false)
-    }
-
-    companion object {
-        @JvmStatic
-        fun newInstance() = WorkoutListFragment()
-    }
-}
