+++ app/src/main/java/com/example/fitnessconstructor/ui/MainActivity.kt
@@ -3,7 +3,9 @@ package com.example.fitnessconstructor.ui
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.example.fitnessconstructor.databinding.ActivityMainBinding
+import dagger.hilt.android.AndroidEntryPoint
 
+@AndroidEntryPoint
 class MainActivity : AppCompatActivity() {
 
     private lateinit var binding: ActivityMainBinding
@@ -12,5 +14,8 @@ class MainActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         binding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
+
+        //remove action bar while launch app
+        supportActionBar?.hide()
     }
 }
