+++ app/src/main/java/com/example/fitnessconstructor/ui/settings/AppSettingsFragment.kt
@@ -1,0 +1,16 @@
+package com.example.fitnessconstructor.ui.settings
+
+import com.example.fitnessconstructor.databinding.FragmentAppSettingsBinding
+import com.example.fitnessconstructor.ui.BaseFragment
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class AppSettingsFragment :
+    BaseFragment<FragmentAppSettingsBinding>(FragmentAppSettingsBinding::inflate) {
+
+    companion object {
+
+        @JvmStatic
+        fun newInstance() = AppSettingsFragment()
+    }
+}
