+++ app/src/main/res/layout/fragment_splash.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SplashFragment"
+    tools:context=".ui.SplashFragment"
     android:background="@drawable/ic_splash_screen">
