+++ app/src/main/res/layout/fragment_statistic.xml
@@ -3,4 +3,4 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".StatisticFragment"/>
+    tools:context=".ui.statistic.StatisticFragment"/>
