+++ app/src/main/res/layout/fragment_workout.xml
@@ -3,4 +3,4 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".WorkoutFragment"/>
+    tools:context=".ui.workout.WorkoutFragment"/>
